+++ src/kli.Localize.Generator/Internal/Helper/DiagnosticsExtensions.cs
@@ -1,11 +1,12 @@
 ﻿using System;
+using System.IO;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.Text;
 using Newtonsoft.Json;
 
-namespace kli.Localize.Generator.Internal.Json
+namespace kli.Localize.Generator.Internal.Helper
 {
-    internal static class JsonDiagnostics
+    internal static class DiagnosticsExtensions
     {
         public static void ReportInvalidFileFormat(this Action<Diagnostic> reporter, string filePath, JsonReaderException ex)
         {
@@ -13,6 +14,10 @@ namespace kli.Localize.Generator.Internal.Json
                 new LinePosition(ex.LineNumber-1, ex.LinePosition), DiagnosticSeverity.Error);
         }
         
+        public static void ReportMissingNeutralCulture(Action<Diagnostic> reporter, AdditionalText additionalText) 
+            => reporter.Report(4, $"Localize element is missing NeutralCulture attribute for files: \"{Path.GetFileName(additionalText.Path)}\".");
+
+
         public static void ReportInvalidKey(this Action<Diagnostic> reporter, string filePath, JsonTextReader reader)
         {
             var message = $"Json property key must be a valid C# identifier: '{reader.Value}'";
@@ -27,6 +32,13 @@ namespace kli.Localize.Generator.Internal.Json
                 new LinePosition(reader.LineNumber-1, reader.LinePosition), DiagnosticSeverity.Warning);
         }
 
+        private static void Report(this Action<Diagnostic> reporter, int id, string message)
+        {
+            var diagnostic = Diagnostic.Create(new DiagnosticDescriptor($"SGL000{id}", "kli.Localize.Generator",
+                message, "Source Generators",
+                DiagnosticSeverity.Error, true), Location.None);
+            reporter(diagnostic);
+        }
         private static void Report(this Action<Diagnostic> reporter, int id, string message,
             string filePath, LinePosition linePosition, DiagnosticSeverity severity)
         {